+++ InverterMatrixFilterElement.cs
@@ -72,7 +72,7 @@ namespace MetaphysicsIndustries.Amethyst
             {
             }
 
-            public override double Modulate(double x)
+            public override float Modulate(float x)
             {
                 return 1 - x;
             }
